Refactoring menu
